HBASE-21652 Refactor ThriftServer making thrift2 server inherited from thrift1 server
