+++ hbase-thrift/src/main/java/org/apache/hadoop/hbase/thrift/HbaseHandlerMetricsProxy.java
@@ -25,9 +25,8 @@ import java.lang.reflect.Proxy;
 
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.hbase.thrift.generated.Hbase;
+import org.apache.hadoop.hbase.thrift2.generated.THBaseService;
 import org.apache.yetus.audience.InterfaceAudience;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 /**
  * Converts a Hbase.Iface using InvocationHandler so that it reports process
@@ -36,10 +35,7 @@ import org.slf4j.LoggerFactory;
 @InterfaceAudience.Private
 public final class HbaseHandlerMetricsProxy implements InvocationHandler {
 
-  private static final Logger LOG = LoggerFactory.getLogger(
-      HbaseHandlerMetricsProxy.class);
-
-  private final Hbase.Iface handler;
+  private final Object handler;
   private final ThriftMetrics metrics;
 
   public static Hbase.Iface newInstance(Hbase.Iface handler,
@@ -51,8 +47,18 @@ public final class HbaseHandlerMetricsProxy implements InvocationHandler {
         new HbaseHandlerMetricsProxy(handler, metrics, conf));
   }
 
+  // for thrift 2
+  public static THBaseService.Iface newInstance(THBaseService.Iface handler,
+      ThriftMetrics metrics,
+      Configuration conf) {
+    return (THBaseService.Iface) Proxy.newProxyInstance(
+        handler.getClass().getClassLoader(),
+        new Class[]{THBaseService.Iface.class},
+        new HbaseHandlerMetricsProxy(handler, metrics, conf));
+  }
+
   private HbaseHandlerMetricsProxy(
-      Hbase.Iface handler, ThriftMetrics metrics, Configuration conf) {
+      Object handler, ThriftMetrics metrics, Configuration conf) {
     this.handler = handler;
     this.metrics = metrics;
   }
